refactor: Extract shared USB audio driver to icom_lan.audio (#376)
Move UsbAudioDriver from icom7610-specific location to shared audio/ package.

**Structural changes:**
- audio.py → audio/lan_stream.py (LAN audio, IC-7610 UDP)
- audio/__init__.py: re-exports all symbols from both modules
- audio/usb_driver.py: UsbAudioDriver (moved from icom7610/drivers/)
- icom7610/drivers/usb_audio.py: backward-compat re-exports only

**Import updates (5 sites):**
- ic705/serial.py → from ...audio.usb_driver import UsbAudioDriver
- ic7300/serial.py → same
- ic9700/serial.py → same
- icom7610/serial.py → same
- cli.py → from .audio.usb_driver import list_usb_audio_devices

**Generalized device patterns:**
Before: ('ic-7610', 'ic 7610', 'icom', 'usb audio codec', 'usb audio')
After: ('usb audio codec', 'usb audio', 'icom', 'yaesu', 'kenwood', 'ftdi')

**Backward compatibility:**
- All old import paths still work via re-exports
- icom_lan.audio.AudioPacket etc still importable
- icom_lan.backends.icom7610.drivers.usb_audio still works
- Zero test regressions (57 sync passed, async failures pre-existing)

**What this enables:**
- YaesuCatRadio can now import UsbAudioDriver from shared location
- Any future radio backend (Kenwood, etc.) uses same driver
- One USB audio implementation for all radios

Next: #377 (integrate AudioCapable into YaesuCatRadio)
